Added policy for Event.

+++ app/Http/Controllers/EventController.php
@@ -26,6 +26,7 @@ class EventController extends Controller
      */
     public function create()
     {
+        $this->authorize("create", Event::class);
         return view("event.create", ['events' => Event::all(), 'rooms' => Room::all(), 'users' => User::all()]);
     }
 
@@ -34,10 +35,10 @@ class EventController extends Controller
      */
     public function store(StoreEventRequest $request)
     {
+        $this->authorize("create", Event::class);
         $user = User::find(Auth::id());
         $form = $request->validated();
-
-        $form["user_id"] = ($user->isAdmin()) ? $request->user_id : Auth::id();
+        $form["user_id"] = ($user->isAdmin()) ? $request->user_id : $user->id;
         if ($request->hasFile('image')) {
             $form['image'] = $request->file('image')->store('images', 'public');
         }
@@ -61,6 +62,7 @@ class EventController extends Controller
      */
     public function edit(Event $event)
     {
+        $this->authorize("update", Event::class);
         return view('event.edit', ['event' => $event, 'events' => Event::all(), 'rooms' => Room::all(), 'users' => User::all()]);
     }
 
@@ -69,10 +71,10 @@ class EventController extends Controller
      */
     public function update(UpdateEventRequest $request, Event $event)
     {
+        $this->authorize("update", Event::class);
         $user = User::find(Auth::id());
         $form = $request->validated();
-
-        $form["user_id"] = ($user->isAdmin()) ? $request->user_id : Auth::id();
+        $form["user_id"] = ($user->isAdmin()) ? $request->user_id : $user->id;
         if ($request->hasFile('image')) {
             $form['image'] = $request->file('image')->store('images', 'public');
         }
@@ -86,6 +88,7 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
+        $this->authorize("delete", Event::class);
         $event->delete();
         return redirect("/events")->with("success", "You have deleted an event");
     }
